Add parallax example
This demonstration has two images, a background and foreground.
The parallax effect is achieved by scrolling both images are different speeds.
Unfortunately, both images are scrolling at the same speed now.
Also, instead of moving the image, we can move the viewport using
Gosu's translate method. This is handy, so check the documentation.